requirements.txt
alembic==1.12.0
bcrypt==3.2.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Cors==4.0.0
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2==2.9.1
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.8.0
urllib3==2.0.4
Werkzeug==2.3.7
WTForms==3.0.1
